refactoring according tasks

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -2,9 +2,6 @@ package ru.kata.spring.boot_security.demo.service;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,18 +10,17 @@ import ru.kata.spring.boot_security.demo.model.User;
 import ru.kata.spring.boot_security.demo.repository.RoleRepo;
 import ru.kata.spring.boot_security.demo.repository.UserRepo;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
+import javax.persistence.EntityExistsException;
+import javax.persistence.EntityNotFoundException;
 import java.util.*;
 import java.util.stream.Collectors;
 
 
 @Service
-public class UserService implements UserDetailsService {
+@Transactional(readOnly = true)
+public class UserService {
 
 
-    @PersistenceContext
-    private EntityManager em;
     private final UserRepo userRepo;
     private final RoleRepo roleRepo;
     private final PasswordEncoder passwordEncoder;
@@ -36,18 +32,9 @@ public class UserService implements UserDetailsService {
         this.passwordEncoder = passwordEncoder;
     }
 
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        User user = userRepo.findUserByEmail(email);
-        if (user == null) {
-            throw new UsernameNotFoundException(String.format("User with '%s' not found", email));
-        }
-        return user;
-    }
 
     // Role methods
-    @Transactional
+
     public List<Role> findAll() {
 
         return roleRepo.findAll();
@@ -61,13 +48,12 @@ public class UserService implements UserDetailsService {
 
     // User methods
 
-    @Transactional
+
     public User findUserByEmail(String email) {
         return userRepo.findUserByEmail(email);
     }
 
 
-    @Transactional
     public List<User> findAllUsers() {
         return userRepo.findAll()
                 .stream()
@@ -78,8 +64,9 @@ public class UserService implements UserDetailsService {
 
     }
 
+
     @Transactional
-    public User addNewUser(User user) throws RuntimeException {
+    public User addNewUser(User user) {
 
         if ((userRepo.findUserByEmail(user.getEmail())) == null) {
             Set<Role> rolesFinal = user.getRole().stream()
@@ -93,7 +80,7 @@ public class UserService implements UserDetailsService {
             userRepo.save(user);
             return user;
         } else {
-            throw new RuntimeException("User is already exists!!!");
+            throw new EntityExistsException("User is already exists!!!");
         }
 
 
@@ -102,6 +89,10 @@ public class UserService implements UserDetailsService {
 
     @Transactional
     public void deleteById(Long id) {
+
+        if (userRepo.findUsersById(id) == null) {
+            throw new EntityNotFoundException("User is not found!!!");
+        }
         userRepo.deleteById(id);
 
 
@@ -110,33 +101,46 @@ public class UserService implements UserDetailsService {
 
     @Transactional
     public User findUsersById(Long id) {
+
+        if (userRepo.findUsersById(id) == null) {
+            throw new EntityNotFoundException("User is not found!!!");
+        }
+
         return userRepo.findUsersById(id);
     }
 
 
     @Transactional
     public User updateUser(long id, User updatedUser) {
-        User userToBeUpdate = userRepo.findUsersById(id);
+
+        if ((userRepo.findUserByEmail(updatedUser.getEmail())) == null) {
 
 
-        Set<Role> rolesFinal = updatedUser.getRole().stream()
-                .flatMap(role2 -> roleRepo.findAll().stream()
-                        .filter(role1 -> role1.getName().equals(role2.getName()))
-                        .map(role1 -> new Role(role1.getId(), role1.getName())))
-                .collect(Collectors.toSet());
+            User userToBeUpdate = userRepo.findUsersById(id);
 
 
-        userToBeUpdate.setFirstName(updatedUser.getFirstName());
-        userToBeUpdate.setLastName(updatedUser.getLastName());
-        userToBeUpdate.setAge(updatedUser.getAge());
-        userToBeUpdate.setGender(updatedUser.getGender());
-        userToBeUpdate.setEmail(updatedUser.getEmail());
-        userToBeUpdate.setPass(passwordEncoder.encode(updatedUser.getPass()));
-        userToBeUpdate.setRole(rolesFinal);
-        userToBeUpdate.setStatus(updatedUser.getStatus());
+            Set<Role> rolesFinal = updatedUser.getRole().stream()
+                    .flatMap(role2 -> roleRepo.findAll().stream()
+                            .filter(role1 -> role1.getName().equals(role2.getName()))
+                            .map(role1 -> new Role(role1.getId(), role1.getName())))
+                    .collect(Collectors.toSet());
 
 
-        return userToBeUpdate;
+            userToBeUpdate.setFirstName(updatedUser.getFirstName());
+            userToBeUpdate.setLastName(updatedUser.getLastName());
+            userToBeUpdate.setAge(updatedUser.getAge());
+            userToBeUpdate.setGender(updatedUser.getGender());
+            userToBeUpdate.setEmail(updatedUser.getEmail());
+            userToBeUpdate.setPass(passwordEncoder.encode(updatedUser.getPass()));
+            userToBeUpdate.setRole(rolesFinal);
+            userToBeUpdate.setStatus(updatedUser.getStatus());
+            return userToBeUpdate;
+
+        } else {
+            throw new EntityExistsException("User is already exists!!!");
+        }
+
+
     }
